% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_bar_labels}
\alias{add_bar_labels}
\title{Add bar labels to ggplot}
\usage{
add_bar_labels(p, df, stacked, label_size)
}
\arguments{
\item{p}{ggplot object}

\item{df}{Data frame with .y column}

\item{stacked}{Logical, whether bars are stacked}

\item{label_size}{Numeric, size of the labels}
}
\description{
Add bar labels to ggplot
}
