% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_common_theme}
\alias{add_common_theme}
\title{Add common theme elements to ggplot}
\usage{
add_common_theme(
  p,
  title,
  x_label,
  y_label,
  note,
  legend_position,
  legend_direction
)
}
\arguments{
\item{p}{ggplot object}

\item{title}{A title}

\item{x_label}{X axis label}

\item{y_label}{Y axis label}

\item{note}{Caption / Note}

\item{legend_position}{Legend position}

\item{legend_direction}{Legend direction}
}
\description{
Add common theme elements to ggplot
}
