% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_facets}
\alias{add_facets}
\title{Add facets to ggplot}
\usage{
add_facets(p, df, scales, ncol)
}
\arguments{
\item{p}{ggplot object}

\item{df}{A data frame}

\item{scales}{Scales for facets}

\item{ncol}{Number of columns for facets}
}
\description{
Add facets to ggplot
}
