% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bar.R
\name{bar_build_base_plot}
\alias{bar_build_base_plot}
\title{Build base ggplot}
\usage{
bar_build_base_plot(df, legend_title)
}
\arguments{
\item{df}{Data frame with .x, .y, .group columns}

\item{legend_title}{Legend title}
}
\value{
ggplot object
}
\description{
Build base ggplot
}
