% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{bar_chart}
\alias{bar_chart}
\title{Bar Chart Component}
\usage{
bar_chart(
  data,
  x,
  y,
  group = NULL,
  title = NULL,
  x_label = NULL,
  y_label = NULL
)
}
\arguments{
\item{data}{Data frame to be used for the chart}

\item{x}{The x variable}

\item{y}{The y variable}

\item{group}{Grouping variable for fill (optional)}

\item{title}{The chart title (optional)}

\item{x_label}{The x axis label (optional)}

\item{y_label}{The y axis label (optional)}
}
\value{
A ggplot2 object
}
\description{
Bar Chart Component
}
