% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bar.R
\name{bar_labels}
\alias{bar_labels}
\title{Bar Labels}
\usage{
bar_labels(p, df, stacked, label_size)
}
\arguments{
\item{p}{ggplot object}

\item{df}{Data frame with .y column}

\item{stacked}{Logical, whether bars are stacked}

\item{label_size}{Label size}
}
\description{
Bar Labels
}
