% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bar.R
\name{bar_position}
\alias{bar_position}
\title{Bar Position}
\usage{
bar_position(stacked, horizontal)
}
\arguments{
\item{stacked}{Stacked bars}

\item{horizontal}{Horizontal bars}
}
\description{
Bar Position
}
