% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bar.R
\name{bar_summarize_data}
\alias{bar_summarize_data}
\title{Summarize plot data for Bars}
\usage{
bar_summarize_data(df)
}
\arguments{
\item{df}{Data frame with .x, .y, .group, .facet columns}
}
\value{
Summarized data frame
}
\description{
Summarize plot data for Bars
}
