% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleted.R
\name{build_bar_base_plot_del}
\alias{build_bar_base_plot_del}
\title{Build base ggplot}
\usage{
build_bar_base_plot_del(df, legend_title)
}
\arguments{
\item{df}{Data frame with .x, .y, .group columns}

\item{legend_title}{Legend title}
}
\value{
ggplot object
}
\description{
Build base ggplot
}
