% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{ChartBook}
\alias{ChartBook}
\title{Generates paginated chart books of performance over time}
\usage{
ChartBook(SYMBOLS)
}
\arguments{
\item{SYMBOLS:}{list of strings of ticker names}
}
\value{
null: generates pdf
}
\description{
Generates paginated chart books of performance over time
}

