% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{color_picker}
\alias{color_picker}
\title{Get a team-standard color palette}
\usage{
color_picker(palette_name = NULL, n = NULL, fixed_mapping = NULL)
}
\arguments{
\item{palette_name}{Character. Name of the palette (categorical, sequential, divergent, etc.)}

\item{n}{Integer. Number of colors required}

\item{fixed_mapping}{Named vector of colors for specific groups (optional)}
}
\value{
Character vector of hex colors
}
\description{
Get a team-standard color palette
}
\examples{
color_picker("categorical", n = 3)
color_picker(fixed_mapping = c(A = "#B3CDE3", B = "#1F77B4"))
}
