% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_composition_barchart.R
\name{draw_composition_barchart}
\alias{draw_composition_barchart}
\title{Composition barchart}
\usage{
draw_composition_barchart(composition, cat1, categorical_values1)
}
\arguments{
\item{composition}{Tibble. Table showing the composition of a test.}

\item{cat1}{Character. First dimension selected.}

\item{categorical_values1}{Tibble. Categories associated to the first dimension selected.}
}
\value{
A ggplot object ready for rendering.
}
\description{
Function creating barchart showing the composition of a test.
}
\author{
Nicolas Mangin
}
