% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_composition_scatterplot.R
\name{draw_composition_scatterplot}
\alias{draw_composition_scatterplot}
\title{Composition scatterplot}
\usage{
draw_composition_scatterplot(composition)
}
\arguments{
\item{composition}{Tibble. Table showing the composition of a test.}
}
\value{
A ggplot object ready for rendering.
}
\description{
Function creating scatterplot showing the composition of a test.
}
\author{
Nicolas Mangin
}
