% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_correlogram.R
\name{draw_correlogram}
\alias{draw_correlogram}
\title{Correlogram}
\usage{
draw_correlogram(scores)
}
\arguments{
\item{scores}{Tibble. Table showing the different scores associated to various observations.}
}
\value{
A ggplot object ready for rendering.
}
\description{
Function creating a correlogram.
}
\author{
Nicolas Mangin
}
