% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_density_plot.R
\name{draw_density_plot}
\alias{draw_density_plot}
\title{Edit diagrams}
\usage{
draw_density_plot(
  scores,
  slctx,
  slcty,
  slctbkgalpha,
  slctpntalpha,
  slctpntsize
)
}
\arguments{
\item{scores}{Tibble.}

\item{slctx}{Character.}

\item{slcty}{Character.}

\item{slctbkgalpha}{Double.}

\item{slctpntalpha}{Double.}

\item{slctpntsize}{Double.}
}
\value{
A set of graphs.
}
\description{
Module facilitating the quick exploration of bivariate relationships.
}
\author{
Nicolas Mangin
}
