% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_diagram.R
\name{draw_diagram}
\alias{draw_diagram}
\title{Create a diagram}
\usage{
draw_diagram(nodes, relations, moderations, translations, language)
}
\arguments{
\item{nodes}{Tibble. Nodes with their properties.}

\item{relations}{Tibble. Edges with their properties.}

\item{moderations}{Tibble. Moderations with their properties.}

\item{translations}{Tibble. Translations for the nodes labels.}

\item{language}{Character. Language to display.}
}
\value{
A DiagrammeR object ready for rendering.
}
\description{
Function creating a diagram object from tibbles specifying nodes, relations, and moderations.
}
\author{
Nicolas Mangin
}
