% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_grade_distribution.R
\name{draw_grade_distribution}
\alias{draw_grade_distribution}
\title{Display grade distribution}
\usage{
draw_grade_distribution(student_grades, pass, increment, facet)
}
\arguments{
\item{student_grades}{Tibble. Table with each "student" id, the maximum number of "points", and the "grade".}

\item{pass}{Numeric. Value above which students validate.}

\item{increment}{Numeric. Size of breaks.}

\item{facet}{Character. Names of the variable which should be used for faceting}
}
\value{
A ggplot object ready for rendering.
}
\description{
Function drawing the distribution of grades as a barplot.
}
\author{
Nicolas Mangin
}
