% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_nomnet.R
\name{draw_nomnet}
\alias{draw_nomnet}
\title{Draw a diagram based on tables detailing nodes, relationships, and moderations.}
\usage{
draw_nomnet(nodes, relations, moderations)
}
\arguments{
\item{nodes}{Tibble. Nodes with their properties.}

\item{relations}{Tibble. Edges with their properties.}

\item{moderations}{Tibble. Moderations with their properties.}
}
\value{
A DiagrammeR object ready for rendering.
}
\description{
Draw a diagram based on tables detailing nodes, relationships, and moderations.
}
