% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eval_plot_vars}
\alias{eval_plot_vars}
\title{Evaluate plot variables}
\usage{
eval_plot_vars(data, x, y, group = NULL, facet = NULL)
}
\arguments{
\item{data}{Data frame}

\item{x}{X variable}

\item{y}{Y variable}

\item{group}{Grouping variable}

\item{facet}{Faceting variable}
}
\value{
Data frame with evaluated variables
}
\description{
Evaluate plot variables
}
