% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_component.R
\name{geom_component}
\alias{geom_component}
\title{Create a geom component function}
\usage{
geom_component(
  geom_layer,
  summarize_fn = summarize_xy,
  position_fn = NULL,
  label_fn = NULL,
  supports_stack = FALSE
)
}
\arguments{
\item{geom_layer}{A ggplot2 geom layer function}

\item{summarize_fn}{A function to summarize data}

\item{position_fn}{A function to create position adjustments}

\item{label_fn}{A function to add labels}

\item{supports_stack}{Logical, whether the geom supports stacking}
}
\description{
Create a geom component function
}
