% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{GetMax}
\alias{GetMax}
\title{gets max in date range (i.e. one_mo, three_mo, etc)
gld_data is the object data.frame(Cl(GLD))}
\usage{
GetMax(DATAFRAME, DATE)
}
\description{
gets max in date range (i.e. one_mo, three_mo, etc)
gld_data is the object data.frame(Cl(GLD))
}

