% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{GetPriceChange}
\alias{GetPriceChange}
\title{Gets delta in price = (b-a) where START is a, END is b
takes from format lubridate::add_with_rollback, returns format to 3 decimal places}
\usage{
GetPriceChange(END, START, DATAFRAME)
}
\arguments{
\item{newer}{formatted date}

\item{older}{formatted date}
}
\value{
difference in price
}
\description{
Gets delta in price = (b-a) where START is a, END is b
takes from format lubridate::add_with_rollback, returns format to 3 decimal places
}

