% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_diagram.R
\name{initialize_diagram}
\alias{initialize_diagram}
\title{Create a starting diagram}
\usage{
initialize_diagram()
}
\value{
A list of tibbles specifying nodes, relations, and moderations.
}
\description{
Function creating a basic diagram serving as a basis to be adjusted with the edit_diagram module.
}
\author{
Nicolas Mangin
}
