% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_line.R
\name{line_add_last_point_labels}
\alias{line_add_last_point_labels}
\title{Add last point labels}
\usage{
line_add_last_point_labels(p, df, x_quo, group_quo, facet_quo, label_size)
}
\arguments{
\item{p}{A ggplot object}

\item{df}{Data frame}

\item{x_quo}{X aesthetic}

\item{group_quo}{Group aesthetic}

\item{facet_quo}{Facet aesthetic}

\item{label_size}{Label size}
}
\value{
A ggplot object
}
\description{
Add last point labels
}
