% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_line.R
\name{line_add_points}
\alias{line_add_points}
\title{Add points}
\usage{
line_add_points(p, df, group_quo, point_shape, point_size)
}
\arguments{
\item{p}{A ggplot object}

\item{df}{Data frame}

\item{group_quo}{Group aesthetic}

\item{point_shape}{Point shape}

\item{point_size}{Point size}
}
\value{
A ggplot object
}
\description{
Add points
}
