% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_line.R
\name{line_build_base_plot}
\alias{line_build_base_plot}
\title{Build base line plot}
\usage{
line_build_base_plot(df, x_quo, y_quo, group_quo, line_size)
}
\arguments{
\item{df}{Data frame}

\item{x_quo}{X aesthetic}

\item{y_quo}{Y aesthetic}

\item{group_quo}{Group aesthetic}

\item{line_size}{Line size}
}
\value{
A ggplot object
}
\description{
Build base line plot
}
