% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartR.R
\name{line_chart}
\alias{line_chart}
\title{Line Chart Component}
\usage{
line_chart(
  data,
  x,
  y,
  group = NULL,
  facet = NULL,
  title = NULL,
  x_label = NULL,
  y_label = NULL,
  legend_title = NULL,
  legend_position = "right",
  legend_direction = "vertical",
  note = NULL,
  facet_scales = "fixed",
  facet_ncol = NULL,
  show_points = FALSE,
  point_shape = NULL,
  point_size = 3.5,
  line_size = 1,
  show_labels = FALSE,
  label_size = 3
)
}
\arguments{
\item{data}{Data Frame}

\item{x}{The x variable}

\item{y}{The y variable}

\item{group}{A grouping variable}

\item{facet}{A faceting variable}

\item{title}{Plot title}

\item{x_label}{X-axis label}

\item{y_label}{Y-axis label}

\item{legend_title}{Legend title}

\item{legend_position}{Legend position}

\item{legend_direction}{Legend direction}

\item{note}{Caption / Note}

\item{facet_scales}{Scales for facets}

\item{facet_ncol}{Number of columns for facets}

\item{show_points}{Show points on line}

\item{point_shape}{Optional vector of shapes per group}

\item{point_size}{Size of points}

\item{line_size}{Size of line}

\item{show_labels}{Show labels on last point per group}

\item{label_size}{Size of labels}
}
\value{
A ggplot2 object
}
\description{
Line Chart Component
}
