% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleted.R
\name{summarize_bar_data_del}
\alias{summarize_bar_data_del}
\title{Summarize plot data for Bars}
\usage{
summarize_bar_data_del(df)
}
\arguments{
\item{df}{Data frame with .x, .y, .group, .facet columns}
}
\value{
Summarized data frame
}
\description{
Summarize plot data for Bars
}
