% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_xy}
\alias{summarize_xy}
\title{Summarize plot data for X-Y plots}
\usage{
summarize_xy(df, stat = "mean")
}
\arguments{
\item{df}{Data frame with .x, .y, .group, .facet columns}

\item{stat}{Statistic to summarize by: "mean", "sum", "count"}
}
\value{
Summarized data frame
}
\description{
Summarize plot data for X-Y plots
}
