% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{Summary}
\alias{Summary}
\title{Generates a summary matrix of percent and price changes}
\usage{
Summary(SYMBOL, max = ten_yr)
}
\arguments{
\item{SYMBOL:}{string ticker name}
}
\value{
summary_data: matrix of percent and price changes
}
\description{
Generates a summary matrix of percent and price changes
}

