% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_density_plot.R
\name{draw_density_plot}
\alias{draw_density_plot}
\title{Draw a density plot}
\usage{
draw_density_plot(
  scores,
  slctx,
  slcty,
  slctbkgalpha,
  slctpntalpha,
  slctpntsize
)
}
\arguments{
\item{scores}{Tibble.}

\item{slctx}{Character.}

\item{slcty}{Character.}

\item{slctbkgalpha}{Double.}

\item{slctpntalpha}{Double.}

\item{slctpntsize}{Double.}
}
\value{
A ggplot object ready for rendering.
}
\description{
Function drawing the 2D density of students along two dimensions selected by the user.
}
\author{
Nicolas Mangin
}
