% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_nomnet.R
\name{draw_nomnet}
\alias{draw_nomnet}
\title{Draw a diagram based on tables detailing nodes, relationships, and moderations.}
\usage{
draw_nomnet(nodes, relations, moderations)
}
\arguments{
\item{nodes}{Tibble. Nodes with their properties.}

\item{relations}{Tibble. Edges with their properties.}

\item{moderations}{Tibble. Moderations with their properties.}
}
\value{
A DiagrammeR object ready for rendering.
}
\description{
Draw a diagram based on tables detailing nodes, relationships, and moderations.
}
\examples{
library(dplyr)
library(tibble)

nodes <- tibble(
  label = c("CTRL", "JUST", "OCB"),
  shape = c("ellipse", "ellipse", "ellipse"),
  x = c(0, 4, 2),
  y = c(0, 0, 2),
  width = 2,
  height = 0.75,
  penwidth = 1,
  color = "black",
  fillcolor = "white",
  fontsize = 14,
  fontcolor = "black",
  include = c(TRUE, TRUE, TRUE)
)

relations <- tibble(
  relation = "CTRL_to_JUST",
  source = "CTRL",
  target = "JUST",
  style = "solid",
  color = "black",
  fontcolor = "black",
  fontsize = 10,
  penwidth = 1,
  arrowhead = "normal",
  label = "+",
  include = c(TRUE)
)

moderations <- tibble(
  source = "OCB",
  target = "CTRL_to_JUST",
  style = "solid",
  color = "black",
  fontcolor = "black",
  fontsize = 10,
  penwidth = 1,
  arrowhead = "normal",
  label = "-",
  include = c(TRUE)
)

draw_nomnet(nodes, relations, moderations) \%>\%
  DiagrammeR::render_graph(width = "100\%")
}
