% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_score_differences.R
\name{draw_score_differences}
\alias{draw_score_differences}
\title{Edit diagrams}
\usage{
draw_score_differences(scores, selection)
}
\arguments{
\item{scores}{Tibble.}

\item{selection}{Character vector}
}
\value{
A ggplot object ready for rendering.
}
\description{
function showing how a set of selected observations differ from the mean of the whole group on a set of questions.
}
\author{
Nicolas Mangin
}
