% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{GetPriceChange}
\alias{GetPriceChange}
\title{Gets price change over period
param START date format
param END date format
param DATAFRAME data frame over period
returns price_change: difference in Cl over period}
\usage{
GetPriceChange(END, START, DATAFRAME)
}
\description{
Gets price change over period
param START date format
param END date format
param DATAFRAME data frame over period
returns price_change: difference in Cl over period
}

