% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{GetPriceChange}
\alias{GetPriceChange}
\title{Gets price change over period}
\usage{
GetPriceChange(DATAFRAME, END, START)
}
\arguments{
\item{DATAFRAME:}{data frame of prices}

\item{START:}{date format}

\item{END:}{date format}
}
\value{
price_change: difference in prices over period
}
\description{
Gets price change over period
}

