% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{line_chart}
\alias{line_chart}
\title{Line Chart Component}
\usage{
line_chart(
  data,
  x,
  y,
  group = NULL,
  title = NULL,
  x_label = NULL,
  y_label = NULL
)
}
\arguments{
\item{data}{Data frame to be used for the chart}

\item{x}{X axis variable}

\item{y}{Y axis variable}

\item{group}{Grouping variable for multiple lines (optional)}

\item{title}{Chart title (optional)}

\item{x_label}{X axis label (optional)}

\item{y_label}{Y axis label (optional)}
}
\value{
A ggplot2 line chart object
}
\description{
Line Chart Component
}
